Support NBT reading/writing on web - Added archive package for compression - Switched Long to BigInt to support js Note: BigInt/Long functionality not fully tested, especially on Bedrock (Little-endian) and  but confirmed working for Java level.dat and player.dat files.

+++ lib/src/nbt/tags/nbt_long.dart
@@ -5,13 +5,13 @@ import 'nbt_tag.dart';
 
 /// Represents a 8 byte long integer in a NBT file.
 class NbtLong extends NbtTag {
-  int _value;
+  BigInt _value;
 
   @override
-  int get value => _value;
+  BigInt get value => _value;
 
   /// Creates a [NbtLong] with given [parent].
-  NbtLong({required String name, required int value})
+  NbtLong({required String name, required BigInt value})
       : _value = value,
         super(name, NbtTagType.TAG_LONG);
